# docs/swagger.yaml
definitions:
...
  models.Course:
    properties:
      description:
        type: string
      duration:
        minimum: 1
        type: integer
      instructor:
        type: string
      price:
        minimum: 0
        type: number
      title:
        type: string
    required:
    - description
    - duration
    - instructor
    - price
    - title
    type: object
...
  models.CreateCourseInput:
    properties:
      description:
        type: string
      duration:
        minimum: 1
        type: integer
      instructor:
        type: string
      price:
        minimum: 0
        type: number
      title:
        type: string
    required:
    - description
    - duration
    - instructor
    - price
    - title
    type: object
  models.CreateProgramInput:
    properties:
      duration:
        type: string
      features:
        items:
          type: string
        minItems: 1
        type: array
      price:
        minimum: 0
        type: number
      title:
        type: string
      type:
        enum:
        - regular
        - intensive
        type: string
    required:
    - duration
    - features
    - price
    - title
    - type
    type: object
...
  models.Program:
    properties:
      duration:
        type: string
      features:
        items:
          type: string
        minItems: 1
        type: array
      price:
        minimum: 0
        type: number
      title:
        type: string
      type:
        enum:
        - regular
        - intensive
        type: string
    required:
    - duration
    - features
    - price
    - title
    - type
    type: object
...
  models.UpdateCourseInput:
    properties:
      description:
        type: string
      duration:
        minimum: 1
        type: integer
      instructor:
        type: string
      price:
        minimum: 0
        type: number
      title:
        type: string
    type: object
  models.UpdateProgramInput:
    properties:
      duration:
        type: string
      features:
        items:
          type: string
        minItems: 1
        type: array
      price:
        minimum: 0
        type: number
      title:
        type: string
      type:
        enum:
        - regular
        - intensive
        type: string
    type: object
  responses.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /courses:
    get:
      consumes:
      - application/json
      description: Retrieve all courses from the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Course'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all courses
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Create a new course with the provided details
      parameters:
      - description: Course creation details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateCourseInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Course'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new course
      tags:
      - courses
  /courses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a course by its ID
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a course
      tags:
      - courses
    get:
      consumes:
      - application/json
      description: Retrieve a specific course by its ID
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Course'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a course by ID
      tags:
      - courses
    put:
      consumes:
      - application/json
      description: Update an existing course's details
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: Course update details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCourseInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Course'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - ApiKeyAuth: []
      summary: Update a course
      tags:
      - courses
...
  /programs:
    get:
      consumes:
      - application/json
      description: Retrieve all programs from the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Program'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all programs
      tags:
      - programs
    post:
      consumes:
      - application/json
      description: Create a new program with the provided details
      parameters:
      - description: Program creation details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateProgramInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Program'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new program
      tags:
      - programs
  /programs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a program by its ID
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a program
      tags:
      - programs
    get:
      consumes:
      - application/json
      description: Retrieve a specific program by its ID
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Program'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a program by ID
      tags:
      - programs
    put:
      consumes:
      - application/json
      description: Update an existing program's details
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      - description: Program update details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProgramInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Program'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - ApiKeyAuth: []
      summary: Update a program
      tags:
      - programs
swagger: "2.0"